refactor: extract GenerateReportDialog shared component
Dialog with notes textarea was duplicated between ReportsPage and
LatestReportPreview. Extracted to GenerateReportDialog with props
for open state, hasExistingReport, and optional onError callback.
