+++ packages/frontend/src/components/reports/ReportsPage.tsx
@@ -1,24 +1,12 @@
 import { useState } from 'react';
 import { RefreshCw, Loader2 } from 'lucide-react';
-import { toast } from 'sonner';
 import type { ApiError } from '@vitals/shared';
+import { toast } from 'sonner';
 import { useReports, useGenerateReport } from '@/api/hooks/useReports';
 import { ReportCard } from './ReportCard';
+import { GenerateReportDialog } from './GenerateReportDialog';
 import { CardSkeleton } from '@/components/ui/LoadingSkeleton';
 import { Button } from '@/components/ui/button';
-import { Textarea } from '@/components/ui/textarea';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-
-export const NOTES_LABEL = 'Notes for AI (optional)';
-export const NOTES_PLACEHOLDER =
-  "Add any context for your report — goals you're tracking, injuries, diet changes, or anything the AI should consider when analyzing your data.";
 
 export function ReportsPage() {
   const { data, isLoading, error } = useReports();
@@ -27,38 +15,18 @@ export function ReportsPage() {
   const hasReports = reports.length > 0;
 
   const [confirmOpen, setConfirmOpen] = useState(false);
-  const [userNotes, setUserNotes] = useState('');
 
-  const handleGenerate = () => {
-    const notes = userNotes.trim();
-    generateReport.mutate(notes ? { userNotes: notes } : undefined, {
-      onSuccess: () => {
-        toast.success('Report generated successfully');
-        handleOpenChange(false);
-      },
-      onError: (err: unknown) => {
-        const status = (err as Partial<ApiError>)?.statusCode;
-        if (status === 429) {
-          toast.error('AI service is rate limited. Please try again in a few minutes.');
-        } else if (status === 502) {
-          toast.error('AI service is temporarily unavailable. Please try again later.');
-        } else if (status === 503) {
-          toast.error('AI service is not configured. Set AI_API_KEY and AI_PROVIDER.');
-        } else {
-          toast.error('Failed to generate report.');
-        }
-        handleOpenChange(false);
-      },
-    });
-  };
-
-  const handleClick = () => {
-    setConfirmOpen(true);
-  };
-
-  const handleOpenChange = (open: boolean) => {
-    setConfirmOpen(open);
-    if (!open) setUserNotes('');
+  const handleError = (err: unknown) => {
+    const status = (err as Partial<ApiError>)?.statusCode;
+    if (status === 429) {
+      toast.error('AI service is rate limited. Please try again in a few minutes.');
+    } else if (status === 502) {
+      toast.error('AI service is temporarily unavailable. Please try again later.');
+    } else if (status === 503) {
+      toast.error('AI service is not configured. Set AI_API_KEY and AI_PROVIDER.');
+    } else {
+      toast.error('Failed to generate report.');
+    }
   };
 
   return (
@@ -68,7 +36,7 @@ export function ReportsPage() {
         <Button
           variant="ghost"
           size="icon-sm"
-          onClick={handleClick}
+          onClick={() => setConfirmOpen(true)}
           disabled={generateReport.isPending}
           title={hasReports ? 'Re-Generate Latest Insights' : 'Generate Latest Insights'}
         >
@@ -102,47 +70,12 @@ export function ReportsPage() {
         )}
       </div>
 
-      <Dialog open={confirmOpen} onOpenChange={handleOpenChange}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>{hasReports ? 'Re-Generate Report?' : 'Generate Report'}</DialogTitle>
-            <DialogDescription>
-              {hasReports
-                ? 'This will generate a new report for the last 7 days, replacing the most recent one.'
-                : 'Generate a new report analyzing your health data from the last 7 days.'}
-            </DialogDescription>
-          </DialogHeader>
-          <div className="space-y-2">
-            <label htmlFor="user-notes" className="text-sm font-medium">
-              {NOTES_LABEL}
-            </label>
-            <Textarea
-              id="user-notes"
-              value={userNotes}
-              onChange={(e) => setUserNotes(e.target.value)}
-              placeholder={NOTES_PLACEHOLDER}
-              rows={4}
-            />
-          </div>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => handleOpenChange(false)}>
-              Cancel
-            </Button>
-            <Button onClick={handleGenerate} disabled={generateReport.isPending}>
-              {generateReport.isPending ? (
-                <>
-                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                  Generating&hellip;
-                </>
-              ) : hasReports ? (
-                'Re-Generate'
-              ) : (
-                'Generate'
-              )}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <GenerateReportDialog
+        open={confirmOpen}
+        onOpenChange={setConfirmOpen}
+        hasExistingReport={hasReports}
+        onError={handleError}
+      />
     </div>
   );
 }
